fix: id값 Long으로 수정 video_id는 건들지 않고 userId값만 건들였습니다.

+++ meerkatai/src/main/java/com/capstone/meerkatai/user/service/AuthServiceImpl.java
@@ -121,7 +121,7 @@ public class AuthServiceImpl implements AuthService {
    */
   @Override
   @Transactional(readOnly = true)
-  public UserInfoResponse getUserInfo(Integer userId) {
+  public UserInfoResponse getUserInfo(Long userId) {
     User user = userRepository.findById(userId)
         .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 사용자입니다."));
